+++ changelog.md
@@ -4,6 +4,7 @@ v1.7.0
 * Added Void Beacon, to conjure something from (almost) nothing
 * Added Cleansing Charm, to help you with your Warp problem
 * Added Bone Bow, revisiting a blast from the past
+* Added Primal Arrows, to spice up your archery
 
 v1.6.0
 ------